+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 1. [#276](https://github.com/influxdata/influxdb-client-js/pull/276): Allow to notify about write success.
 1. [#280](https://github.com/influxdata/influxdb-client-js/pull/280): Upgrade to the latest node v14 LTS.
 1. [#283](https://github.com/influxdata/influxdb-client-js/pull/283): Allow to specify custom HTTP headers in WriteApi or QueryApi.
+1. [#287](https://github.com/influxdata/influxdb-client-js/pull/287): Throw HttpError with code and message from response body.
 
 ## 1.8.0 [2020-10-30]